enable joining and leaving of a game room

+++ src/containers/Game/Game.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { useParams } from 'react-router-dom';
+import { useParams, useHistory } from 'react-router-dom';
 import Loader from '../../components/Loader';
+import * as ROUTES from '../../constants/routes';
 import { useAuth } from '../../store/auth';
 import { useFirebase } from '../../store/firebase';
 
@@ -8,23 +9,52 @@ const Game = () => {
   const [loading, setLoading] = useState(true);
   const [game, setGame] = useState(null);
   const [isHost, setHost] = useState(false);
+  const [isChallenger, setChallenger] = useState(false);
 
   const { id } = useParams();
+  const history = useHistory();
   const auth = useAuth();
   const authRef = useRef(auth);
   const firebase = useFirebase();
   const firebaseRef = useRef(firebase);
 
-  const deleteGame = useCallback(() => {
-    firebaseRef.current.game(id).delete();
-  }, [id]);
+  const onLeave = useCallback(() => {
+    if (isHost) {
+      firebaseRef.current.game(id).delete();
+    }
+
+    if (isChallenger) {
+      firebaseRef.current.game(id).update({
+        challenger: null,
+      });
+    }
+  }, [id, isHost, isChallenger]);
 
   useEffect(() => {
     const listener = firebaseRef.current.game(id).onSnapshot((snapshot) => {
       const gameData = snapshot.data();
 
+      // if game does not exist, bounce to home
+      if (!gameData) {
+        history.push(ROUTES.LANDING);
+        return;
+      }
+
+      // determine if current user is the host
       if (gameData.host.id === authRef.current.user.id) {
         setHost(true);
+      } else if (!gameData.challenger) {
+        // if no challenger, set this user as challenger
+        firebaseRef.current.game(id).update({
+          challenger: authRef.current.user,
+        });
+        setChallenger(true);
+      } else if (
+        gameData.challenger &&
+        gameData.challenger.id !== authRef.current.user.id
+      ) {
+        // if challenger already exists
+        history.push(ROUTES.GAMES);
       }
 
       setGame(gameData);
@@ -33,13 +63,10 @@ const Game = () => {
     setLoading(false);
 
     return () => {
-      if (isHost) {
-        deleteGame();
-      }
-
       listener(); // unsubscribe to listener
+      onLeave(); // clean up the room
     };
-  }, [id, isHost, deleteGame]);
+  }, [id, history, onLeave]);
 
   if (loading) {
     return <Loader />;
